Remove created files during build
Currently, make clean doesn't return to the state the build was before the make
was called, and this is wrong. The state of the files after 'make clean'
and 'make distclean' should be the same as before calling 'make'.

At the moment, I am removing the config logs and status, as a created binary
during 'make check'